updating the module folder
