+++ braindecode/modules/attention.py
@@ -15,6 +15,7 @@ import math
 
 import torch
 import torch.nn.functional as F
+
 from einops import rearrange
 from einops.layers.torch import Rearrange
 from torch import Tensor, nn
